[REFACTOR] Facade 패키지 Domain 하위로 리패키징
