+++ src/main/java/kr/hhplus/be/server/api/point/controller/PointController.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.api.point.controller;
 
 import kr.hhplus.be.server.api.point.controller.request.PointAddRequest;
 import kr.hhplus.be.server.api.point.controller.response.PointResponse;
-import kr.hhplus.be.server.api.point.facade.PointFacade;
+import kr.hhplus.be.server.domain.point.facade.PointFacade;
 import kr.hhplus.be.server.domain.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
