+++ src/test/java/kr/hhplus/be/server/integration/api/product/facade/ProductFacadeTest.java
@@ -2,13 +2,13 @@ package kr.hhplus.be.server.integration.api.product.facade;
 
 import kr.hhplus.be.server.api.order.controller.request.OrderCreateRequest;
 import kr.hhplus.be.server.api.order.controller.request.OrderProductDetail;
-import kr.hhplus.be.server.api.order.facade.OrderFacade;
+import kr.hhplus.be.server.domain.order.facade.OrderFacade;
 import kr.hhplus.be.server.api.payment.controller.request.PaymentRequest;
-import kr.hhplus.be.server.api.payment.facade.PaymentFacade;
+import kr.hhplus.be.server.domain.payment.facade.PaymentFacade;
 import kr.hhplus.be.server.api.product.controller.response.BestSellingProductResponse;
 import kr.hhplus.be.server.api.product.controller.response.ProductAllResponse;
 import kr.hhplus.be.server.api.product.controller.response.ProductResponse;
-import kr.hhplus.be.server.api.product.facade.ProductFacade;
+import kr.hhplus.be.server.domain.product.facade.ProductFacade;
 import kr.hhplus.be.server.domain.payment.domain.PaymentMethod;
 import kr.hhplus.be.server.domain.point.domain.Point;
 import kr.hhplus.be.server.domain.point.repository.PointRepository;
