+++ src/test/java/kr/hhplus/be/server/unit/api/point/facade/PointFacadeTest.java
@@ -2,7 +2,7 @@ package kr.hhplus.be.server.unit.api.point.facade;
 
 import kr.hhplus.be.server.api.point.controller.request.PointAddRequest;
 import kr.hhplus.be.server.api.point.controller.response.PointResponse;
-import kr.hhplus.be.server.api.point.facade.PointFacade;
+import kr.hhplus.be.server.domain.point.facade.PointFacade;
 import kr.hhplus.be.server.domain.point.domain.Point;
 import kr.hhplus.be.server.domain.point.domain.PointTransaction;
 import kr.hhplus.be.server.domain.point.domain.PointTransactionType;
